+++ src/resolvers/User/likes/likes.graphql
@@ -1,0 +1,3 @@
+type User {
+  likes: Mylist
+}
